requirements.txt
tensorflow[and-cuda]==2.19.1
keras==3.6.0
opencv-python==4.10.0.84
kaggle>=1.6
streamlit>=1.57
jupyter>=1.1
ipykernel>=6.29
numpy>=2.1
pandas>=3.0
matplotlib>=3.10
seaborn>=0.13
scikit-learn>=1.8
pillow>=12.0
h5py>=3.14
tqdm>=4.67
